+++ lib/wasi-types/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "wasmer-wasi-types"
+version = "2.0.0-rc2"
+description = "WASI types for Wasmer WebAssembly runtime"
+categories = ["wasm", "os"]
+keywords = ["wasm", "webassembly", "wasi", "sandbox", "ABI"]
+authors = ["Wasmer Engineering Team <engineering@wasmer.io>"]
+repository = "https://github.com/wasmerio/wasmer"
+license = "MIT"
+readme = "README.md"
+edition = "2018"
+
+[dependencies]
+wasmer = { path = "../api", version = "2.0.0-rc2" }
+serde = { version = "1.0", features = ["derive"] }
+byteorder = "1.3"
+time = "0.1"
