+++ lib/wasi-types/src/time.rs
@@ -1,0 +1,7 @@
+pub type __wasi_clockid_t = u32;
+pub const __WASI_CLOCK_REALTIME: u32 = 0;
+pub const __WASI_CLOCK_MONOTONIC: u32 = 1;
+pub const __WASI_CLOCK_PROCESS_CPUTIME_ID: u32 = 2;
+pub const __WASI_CLOCK_THREAD_CPUTIME_ID: u32 = 3;
+
+pub type __wasi_timestamp_t = u64;
