+++ lib/wasi-types/src/versions/mod.rs
@@ -1,0 +1,1 @@
+pub mod snapshot0;
