+++ lib/wasi-types/src/versions/snapshot0.rs
@@ -1,0 +1,101 @@
+use crate::*;
+use serde::{Deserialize, Serialize};
+use std::fmt;
+use wasmer::ValueType;
+
+pub type __wasi_linkcount_t = u32;
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+#[repr(C)]
+pub struct __wasi_subscription_clock_t {
+    pub userdata: __wasi_userdata_t,
+    pub clock_id: __wasi_clockid_t,
+    pub timeout: __wasi_timestamp_t,
+    pub precision: __wasi_timestamp_t,
+    pub flags: __wasi_subclockflags_t,
+}
+
+#[derive(Copy, Clone)]
+#[repr(C)]
+pub union __wasi_subscription_u {
+    pub clock: __wasi_subscription_clock_t,
+    pub fd_readwrite: __wasi_subscription_fs_readwrite_t,
+}
+
+#[derive(Copy, Clone)]
+#[repr(C)]
+pub struct __wasi_subscription_t {
+    pub userdata: __wasi_userdata_t,
+    pub type_: __wasi_eventtype_t,
+    pub u: __wasi_subscription_u,
+}
+
+unsafe impl ValueType for __wasi_subscription_t {}
+
+pub type __wasi_whence_t = u8;
+pub const __WASI_WHENCE_CUR: u8 = 0;
+pub const __WASI_WHENCE_END: u8 = 1;
+pub const __WASI_WHENCE_SET: u8 = 2;
+
+#[derive(Copy, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[repr(C)]
+pub struct __wasi_filestat_t {
+    pub st_dev: __wasi_device_t,
+    pub st_ino: __wasi_inode_t,
+    pub st_filetype: __wasi_filetype_t,
+    pub st_nlink: __wasi_linkcount_t,
+    pub st_size: __wasi_filesize_t,
+    pub st_atim: __wasi_timestamp_t,
+    pub st_mtim: __wasi_timestamp_t,
+    pub st_ctim: __wasi_timestamp_t,
+}
+
+unsafe impl ValueType for __wasi_filestat_t {}
+
+impl fmt::Debug for __wasi_filestat_t {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let convert_ts_into_time_string = |ts| {
+            let tspec =
+                ::time::Timespec::new(ts as i64 / 1_000_000_000, (ts % 1_000_000_000) as i32);
+            let tm = ::time::at(tspec);
+            let out_time = tm.rfc822();
+            format!("{} ({})", out_time, ts)
+        };
+        f.debug_struct("__wasi_filestat_t")
+            .field("st_dev", &self.st_dev)
+            .field("st_ino", &self.st_ino)
+            .field(
+                "st_filetype",
+                &format!(
+                    "{} ({})",
+                    wasi_filetype_to_name(self.st_filetype),
+                    self.st_filetype,
+                ),
+            )
+            .field("st_nlink", &self.st_nlink)
+            .field("st_size", &self.st_size)
+            .field("st_atim", &convert_ts_into_time_string(self.st_atim))
+            .field("st_mtim", &convert_ts_into_time_string(self.st_mtim))
+            .field("st_ctim", &convert_ts_into_time_string(self.st_ctim))
+            .finish()
+    }
+}
+
+impl fmt::Debug for __wasi_subscription_t {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("__wasi_subscription_t")
+            .field("userdata", &self.userdata)
+            .field("type", &eventtype_to_str(self.type_))
+            .field(
+                "u",
+                match self.type_ {
+                    __WASI_EVENTTYPE_CLOCK => unsafe { &self.u.clock },
+                    __WASI_EVENTTYPE_FD_READ | __WASI_EVENTTYPE_FD_WRITE => unsafe {
+                        &self.u.fd_readwrite
+                    },
+                    _ => &"INVALID EVENTTYPE",
+                },
+            )
+            .finish()
+    }
+}
