+++ lib/wasi/Cargo.toml
@@ -12,15 +12,14 @@ edition = "2018"
 
 [dependencies]
 bincode = "1"
-byteorder = "1.3"
 thiserror = "1"
 generational-arena = { version = "0.2", features = ["serde"] }
 libc = { version = "^0.2", default-features = false }
 tracing = { version = "0.1" }
 getrandom = "0.2"
-time = "0.1"
 typetag = "0.1"
 serde = { version = "1.0", features = ["derive"] }
+wasmer-wasi-types = { path = "../wasi-types", version = "2.0.0-rc2" }
 wasmer = { path = "../api", version = "2.0.0-rc2", default-features = false }
 
 [target.'cfg(windows)'.dependencies]
